--- app/src/main/java/io/intelehealth/client/offline_login/NetworkConnection.java
@@ -1,21 +1,0 @@
-package io.intelehealth.client.offline_login;
-
-import android.content.Context;
-import android.net.ConnectivityManager;
-import android.net.NetworkInfo;
-
-/**
- * Created by Dexter Barretto on 5/24/17.
- * Github : @dbarretto
- */
-
-public class NetworkConnection {
-
-    public static boolean isOnline(Context context) {
-        ConnectivityManager cm =
-                (ConnectivityManager) context.getSystemService(Context.CONNECTIVITY_SERVICE);
-        NetworkInfo netInfo = cm.getActiveNetworkInfo();
-        return netInfo != null && netInfo.isConnected();
-    }
-
-}
