+++ app/src/main/java/io/intelehealth/client/utils/NetworkConnection.java
@@ -1,0 +1,37 @@
+package io.intelehealth.client.utils;
+
+import android.content.Context;
+import android.net.ConnectivityManager;
+import android.net.NetworkInfo;
+
+/**
+ * Created by Dexter Barretto on 5/24/17.
+ * Github : @dbarretto
+ */
+
+public class NetworkConnection {
+
+    public static final String TYPE_WIFI = "WIFI";
+    public static final String TYPE_MOBILE = "MOBILE_DATA";
+    public static final String TYPE_NOT_CONNECTED = "NO_CONNECTION";
+
+    public static boolean isOnline(Context context) {
+        ConnectivityManager cm =
+                (ConnectivityManager) context.getSystemService(Context.CONNECTIVITY_SERVICE);
+        NetworkInfo netInfo = cm.getActiveNetworkInfo();
+        return netInfo != null && netInfo.isConnected();
+    }
+
+    public static String getConnectionType(Context context) {
+        ConnectivityManager cm =
+                (ConnectivityManager) context.getSystemService(Context.CONNECTIVITY_SERVICE);
+        NetworkInfo activeNetwork = cm.getActiveNetworkInfo();
+        if (activeNetwork != null) {
+            if (activeNetwork.getType() == ConnectivityManager.TYPE_WIFI) return TYPE_WIFI;
+            if (activeNetwork.getType() == ConnectivityManager.TYPE_MOBILE) return TYPE_MOBILE;
+        }
+        return TYPE_NOT_CONNECTED;
+    }
+
+}
+
